Part 2 of Day 1 as well.
Without reddit help, I wouldn't have been able to complete with part 1.
Even without tricks, part 1 completes fast enough, but part 2 would not.
Another trick that was mentioned and is worth remembering, but not
implemented, is to run two BFS from start and from end, in lockstep,
and stop when they meet.  That might be necessary for longer problems.